modify variable name, add dropout_
