+++ paddle/fluid/pir/dialect/operator/interface/infer_symbolic_shape/binary_infer_sym.h
@@ -44,6 +44,7 @@ OP_DECLARE_INFER_SYMBOLIC_SHAPE(DepthwiseConv2d)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(DepthwiseConv2dTranspose)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Dot)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Dropout)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Dropout_)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Embedding)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(EqualAll)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(SparseWeightEmbedding)
