+++ test/legacy_test/test_dropout_op_inplace.py
@@ -54,17 +54,6 @@ class TestDropoutInplaceDygraph(unittest.TestCase):
                     result.numpy(), np.zeros_like(in_np), rtol=1e-05
                 )
 
-    def test_inplace_axis(self):
-        for place in self.places:
-            with base.dygraph.guard(place):
-                in_np = np.random.random([2, 3, 4]).astype("float32")
-                input = paddle.to_tensor(in_np)
-                input_id = id(input)
-                result = paddle.nn.functional.dropout(
-                    x=input, p=0.0, axis=0, training=True, inplace=True
-                )
-                self.assertEqual(id(result), input_id)
-
     def test_inplace_downscale(self):
         for place in self.places:
             with base.dygraph.guard(place):
@@ -104,64 +93,21 @@ class TestDropoutInplaceAxisDygraph(unittest.TestCase):
         np.random.seed(123)
         self.places = get_places()
 
-    def test_inplace_axis_training(self):
+    def test_inplace_axis_error(self):
         for place in self.places:
-            with base.dygraph.guard(place):
+            with (
+                base.dygraph.guard(place),
+                self.assertRaises(NotImplementedError),
+            ):
                 in_np = np.random.random([2, 3, 4]).astype("float32")
                 input = paddle.to_tensor(in_np)
-                input_id = id(input)
                 result = paddle.nn.functional.dropout(
                     x=input,
                     p=0.5,
                     axis=1,
                     training=True,
-                    mode='upscale_in_train',
                     inplace=True,
                 )
-                self.assertEqual(id(result), input_id)
-
-    def test_inplace_axis_p_one_dynamic(self):
-        for place in self.places:
-            with base.dygraph.guard(place):
-                if not isinstance(place, paddle.CPUPlace):
-                    in_np = np.random.random([2, 3, 4]).astype("float32")
-                    input = paddle.to_tensor(in_np)
-                    input_id = id(input)
-                    result = paddle.nn.functional.dropout(
-                        x=input, p=1.0, axis=0, training=True, inplace=True
-                    )
-                    self.assertEqual(id(result), input_id)
-
-    def test_inplace_axis_p_one_pir(self):
-        with (
-            paddle.pir_utils.IrGuard(),
-            paddle.static.program_guard(
-                paddle.static.Program(), paddle.static.Program()
-            ),
-        ):
-            in_np = np.random.random([2, 3, 4]).astype("float32")
-            input = paddle.to_tensor(in_np)
-            input_id = id(input)
-            result = paddle.nn.functional.dropout(
-                x=input, p=1.0, axis=0, training=True, inplace=True
-            )
-            self.assertEqual(id(result), input_id)
-
-    def test_inplace_downscale_axis(self):
-        for place in self.places:
-            with base.dygraph.guard(place):
-                in_np = np.random.random([2, 3, 4]).astype("float32")
-                input = paddle.to_tensor(in_np)
-                input_id = id(input)
-                result = paddle.nn.functional.dropout(
-                    x=input,
-                    p=0.5,
-                    axis=[0, 1],
-                    mode='downscale_in_infer',
-                    training=False,
-                    inplace=True,
-                )
-                self.assertEqual(id(result), input_id)
 
 
 class TestDropoutLayerInplace(unittest.TestCase):
