+++ test/legacy_test/test_normal.py
@@ -409,120 +409,5 @@ class TestNormalComplexErrors(unittest.TestCase):
             self.assertRaises(ValueError, paddle.normal, mean)
 
 
-class TestNormalAPI_out_parameter(unittest.TestCase):
-    def test_out_with_shape(self):
-        paddle.disable_static()
-        shape = [2, 3]
-
-        out_tensor = paddle.empty(shape, dtype='float32')
-        original_ptr = out_tensor.data_ptr()
-
-        result = paddle.normal(mean=0.0, std=1.0, shape=shape, out=out_tensor)
-
-        self.assertEqual(result.data_ptr(), original_ptr)
-        self.assertEqual(result.data_ptr(), out_tensor.data_ptr())
-        self.assertEqual(list(result.shape), shape)
-
-        paddle.enable_static()
-
-    def test_out_with_mean_tensor(self):
-        paddle.disable_static()
-
-        mean_tensor = paddle.to_tensor([1.0, 2.0, 3.0])
-        shape = [3]
-
-        out_tensor = paddle.empty(shape, dtype='float32')
-        original_ptr = out_tensor.data_ptr()
-
-        result = paddle.normal(mean=mean_tensor, std=1.0, out=out_tensor)
-
-        self.assertEqual(result.data_ptr(), original_ptr)
-        self.assertEqual(result.data_ptr(), out_tensor.data_ptr())
-        self.assertEqual(list(result.shape), shape)
-
-        paddle.enable_static()
-
-    def test_out_with_std_tensor(self):
-        paddle.disable_static()
-
-        std_tensor = paddle.to_tensor([1.0, 2.0, 3.0])
-        shape = [3]
-
-        out_tensor = paddle.empty(shape, dtype='float32')
-        original_ptr = out_tensor.data_ptr()
-
-        result = paddle.normal(mean=0.0, std=std_tensor, out=out_tensor)
-
-        self.assertEqual(result.data_ptr(), original_ptr)
-        self.assertEqual(result.data_ptr(), out_tensor.data_ptr())
-        self.assertEqual(list(result.shape), shape)
-
-        paddle.enable_static()
-
-    def test_out_with_mean_std_tensors(self):
-        paddle.disable_static()
-
-        mean_tensor = paddle.to_tensor([1.0, 2.0, 3.0])
-        std_tensor = paddle.to_tensor([0.5, 1.0, 1.5])
-        shape = [3]
-
-        out_tensor = paddle.empty(shape, dtype='float32')
-        original_ptr = out_tensor.data_ptr()
-
-        result = paddle.normal(mean=mean_tensor, std=std_tensor, out=out_tensor)
-
-        self.assertEqual(result.data_ptr(), original_ptr)
-        self.assertEqual(result.data_ptr(), out_tensor.data_ptr())
-        self.assertEqual(list(result.shape), shape)
-
-        paddle.enable_static()
-
-    def test_out_with_complex_mean(self):
-        paddle.disable_static()
-
-        shape = [2, 3]
-
-        out_tensor = paddle.empty(shape, dtype='complex64')
-        original_ptr = out_tensor.data_ptr()
-
-        result = paddle.normal(
-            mean=1.0 + 1.0j, std=1.0, shape=shape, out=out_tensor
-        )
-
-        self.assertEqual(result.data_ptr(), original_ptr)
-        self.assertEqual(result.data_ptr(), out_tensor.data_ptr())
-        self.assertEqual(list(result.shape), shape)
-        self.assertEqual(result.dtype, paddle.complex64)
-
-        paddle.enable_static()
-
-    def test_out_with_complex_mean_tensor(self):
-        paddle.disable_static()
-
-        mean_tensor = paddle.to_tensor([1.0 + 1.0j, 2.0 + 2.0j, 3.0 + 3.0j])
-        shape = [3]
-
-        out_tensor = paddle.empty(shape, dtype='complex64')
-        original_ptr = out_tensor.data_ptr()
-
-        result = paddle.normal(mean=mean_tensor, std=1.0, out=out_tensor)
-
-        self.assertEqual(result.data_ptr(), original_ptr)
-        self.assertEqual(result.data_ptr(), out_tensor.data_ptr())
-        self.assertEqual(list(result.shape), shape)
-        self.assertEqual(result.dtype, paddle.complex64)
-
-        paddle.enable_static()
-
-
-class TestNormalAPI_size_alias(unittest.TestCase):
-    def test_size_alias_basic(self):
-        paddle.disable_static()
-        shape = [2, 3]
-        out = paddle.normal(mean=0.0, std=1.0, size=shape)
-        self.assertEqual(list(out.shape), shape)
-        paddle.enable_static()
-
-
 if __name__ == "__main__":
     unittest.main()
